% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_whole_gene}
\alias{run_whole_gene}
\title{Run the GAM pipeline for a list of genes}
\usage{
run_whole_gene(
  sce1,
  sce2,
  geneList,
  binned = FALSE,
  bootstrap = TRUE,
  n_bootstrap = 100,
  m_prop = 0.5,
  n_cores = 1
)
}
\arguments{
\item{sce1}{Single-cell experiment object 1 (reference)}

\item{sce2}{Single-cell experiment object 2 (target)}

\item{geneList}{List of gene names}

\item{binned}{Logical, whether to bin the data (default: FALSE)}

\item{bootstrap}{Logical, whether to bootstrap the data (default: TRUE)}

\item{n_bootstrap}{Number of bootstrap samples (default: 100)}

\item{m_prop}{Proportion of data to sample for bootstrap (default: 0.5)}

\item{n_cores}{Number of cores to use for parallel processing (default: 1)}
}
\value{
List of values, arrays, and p-values for each gene
}
\description{
Run the GAM pipeline for a list of genes
}
